update references to aslp-server and aslp-cpp (possibly breaking!) (#116)
we've moved around our repositories to separate the client/server code
from the main aslp repository.

possible breaking! this adds a new parameter to the aslp_connection
constructor. this will require a newer version of aslp-cpp.
if you are building with ./build.sh, deleting the build directory should
automatically fetch the latest version.